Known good state

+++ src/world/world.h
@@ -28,20 +28,20 @@ typedef void block;
 
 #include <stdlib.h>
 #include <stdbool.h>
+#include <ninstd/types.h>
 
 #include <world/positions.h>
 #include "../network/player.h"
 
 struct player; // TODO.. what the hell? super strange bug, why is this required??
 
-int world_manager_init(void);
+i64 world_manager_init(void);
 void world_manager_cleanup(void);
 void world_do_tick(world *world);
 size_t world_manager_get_world_count();
 world **world_manager_get_worlds();
 struct entitypos world_manager_get_init_spawn_pos(void);
 bool world_send_chunk_data1(const struct player *p);
-//void testerino(struct testerino *t);
 
 
 #endif
